sellerweb/requirements.txt
django 
djangorestframework 
djoser 
python-decouple